requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
torch>=2.0.0
scikit-learn>=1.3.0
PuLP>=2.7.0
plotly>=5.14.0
dash>=2.11.0
openpyxl>=3.1.0
python-docx>=0.8.11
matplotlib>=3.7.0
pyyaml>=6.0
opcua>=0.98.13

# API & server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
xlrd>=2.0.0
websockets>=12.0
requests>=2.31.0
python-dotenv>=1.0.0
prometheus-client>=0.17.0

# Chatbot (Gemini)
google-generativeai>=0.3.0

# ---------------------------
# Development & QA tools
# ---------------------------
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0

types-PyYAML>=6.0.0
types-requests>=2.31.0

ipython>=8.14.0
ipdb>=0.13.13
